Move EventHandlerFunction and EventTreeService to 'core.hooks' package
