+++ src/main/kotlin/io/github/freya022/botcommands/internal/core/hooks/EventDispatcherImpl.kt
@@ -6,7 +6,6 @@ import io.github.freya022.botcommands.api.core.annotations.BEventListener.RunMod
 import io.github.freya022.botcommands.api.core.config.BCoroutineScopesConfig
 import io.github.freya022.botcommands.api.core.events.InitializationEvent
 import io.github.freya022.botcommands.api.core.utils.simpleNestedName
-import io.github.freya022.botcommands.internal.core.EventHandlerFunction
 import io.github.freya022.botcommands.internal.utils.shortSignature
 import io.github.freya022.botcommands.internal.utils.unwrap
 import io.github.oshai.kotlinlogging.KotlinLogging
