+++ src/main/kotlin/io/github/freya022/botcommands/internal/core/hooks/EventHooksProvider.kt
@@ -8,7 +8,6 @@ import io.github.freya022.botcommands.api.core.config.BCoroutineScopesConfig
 import io.github.freya022.botcommands.api.core.service.ServiceContainer
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.api.core.service.annotations.ServiceType
-import io.github.freya022.botcommands.internal.core.EventTreeService
 import io.github.freya022.botcommands.internal.core.service.FunctionAnnotationsMap
 import net.dv8tion.jda.api.hooks.IEventManager
 import org.springframework.context.annotation.Bean
